Refactor header and sidebar
