+++ pages/agents/[id].tsx
@@ -4,6 +4,8 @@ import Link from 'next/link';
 import React from 'react';
 
 import Sidebar from '@/components/Sidebar';
+import HamburgerIcon from '@/components/HamburgerIcon';
+import CloseIcon from '@/components/CloseIcon';
 
 
 // Функция для форматирования текста с абзацами
@@ -108,11 +110,6 @@ export default function AgentChat() {
     setSidebarOpen(window.innerWidth > 768);
   }, []);
 
-  useEffect(() => {
-    const handler = () => setSidebarOpen(prev => !prev);
-    document.addEventListener('toggleSidebar', handler);
-    return () => document.removeEventListener('toggleSidebar', handler);
-  }, []);
   
   // Автоматический скролл к последнему сообщению
   const scrollToBottom = () => {
@@ -176,6 +173,17 @@ export default function AgentChat() {
     setUserMenuOpen(!userMenuOpen);
   };
 
+  const handleLogout = async () => {
+    try {
+      const res = await fetch('/api/logout', { credentials: 'include' });
+      if (res.ok) {
+        window.location.href = '/auth/login';
+      }
+    } catch (e) {
+      console.error('Ошибка при выходе:', e);
+    }
+  };
+
   async function sendMessage() {
     if (!input.trim() || typeof id !== 'string') return;
 
@@ -231,6 +239,36 @@ export default function AgentChat() {
 
       {/* Main Chat Content */}
       <main className={`chat-main ${sidebarOpen ? 'with-sidebar' : 'full-width'}`}>
+        <header className="lk-header">
+          <button className="mobile-hamburger" onClick={toggleSidebar}>
+            {sidebarOpen ? <CloseIcon /> : <HamburgerIcon />}
+          </button>
+          <div className="lk-header__profile">
+            <button className="user-info-button" onClick={toggleUserMenu}>
+              <div className="lk-header__avatar">{email.charAt(0).toUpperCase()}</div>
+              <div className="user-details">
+                <div className="user-email">{email}</div>
+                <div className={`subscription-badge ${subscriptionStatus}`}>{
+                  subscriptionStatus === 'trial'
+                    ? 'Пробный период'
+                    : subscriptionStatus === 'active'
+                      ? 'Активна'
+                      : 'Истекла'
+                }</div>
+              </div>
+              <div className="chevron">{userMenuOpen ? '▲' : '▼'}</div>
+            </button>
+            {userMenuOpen && (
+              <div className="user-menu">
+                <Link href="/settings" className="user-menu-item">Настройки</Link>
+                <Link href="/profile" className="user-menu-item">Профиль</Link>
+                <Link href="/help" className="user-menu-item">Помощь</Link>
+                <hr className="user-menu-separator" />
+                <button onClick={handleLogout} className="user-menu-item">Выйти</button>
+              </div>
+            )}
+          </div>
+        </header>
         <div className="chat-header">
           <h1 className="chat-title">Чат с {assistantName}</h1>
           <button
