+++ pages/agents/index.tsx
@@ -2,22 +2,20 @@
 import React, { useEffect, useState } from 'react';
 import Link from 'next/link';
 import Sidebar from '@/components/Sidebar';
+import HamburgerIcon from '@/components/HamburgerIcon';
+import CloseIcon from '@/components/CloseIcon';
 
 export default function AgentsPage() {
   const [email, setEmail] = useState('');
   const [subscriptionStatus, setSubscriptionStatus] = useState<'active' | 'trial' | 'expired'>('trial');
   const [sidebarOpen, setSidebarOpen] = useState(false);
+  const [userMenuOpen, setUserMenuOpen] = useState(false);
   const [loading, setLoading] = useState(true);
 
   useEffect(() => {
     setSidebarOpen(window.innerWidth > 768);
   }, []);
 
-  useEffect(() => {
-    const handler = () => setSidebarOpen(prev => !prev);
-    document.addEventListener('toggleSidebar', handler);
-    return () => document.removeEventListener('toggleSidebar', handler);
-  }, []);
 
   const [categories, setCategories] = useState<any[]>([]);
   const [agents, setAgents] = useState<any[]>([]);
@@ -49,6 +47,18 @@ export default function AgentsPage() {
   }, []);
 
   const toggleSidebar = () => setSidebarOpen(!sidebarOpen);
+  const toggleUserMenu = () => setUserMenuOpen(!userMenuOpen);
+
+  const handleLogout = async () => {
+    try {
+      const res = await fetch('/api/logout', { credentials: 'include' });
+      if (res.ok) {
+        window.location.href = '/auth/login';
+      }
+    } catch (e) {
+      console.error('Ошибка при выходе:', e);
+    }
+  };
 
   if (loading) {
     return <div>Загрузка...</div>;
@@ -66,6 +76,36 @@ export default function AgentsPage() {
       />
 
       <main className={`main-content ${sidebarOpen ? 'with-sidebar' : 'full-width'}`}>
+        <header className="lk-header">
+          <button className="mobile-hamburger" onClick={toggleSidebar}>
+            {sidebarOpen ? <CloseIcon /> : <HamburgerIcon />}
+          </button>
+          <div className="lk-header__profile">
+            <button className="user-info-button" onClick={toggleUserMenu}>
+              <div className="lk-header__avatar">{email.charAt(0).toUpperCase()}</div>
+              <div className="user-details">
+                <div className="user-email">{email}</div>
+                <div className={`subscription-badge ${subscriptionStatus}`}>{
+                  subscriptionStatus === 'trial'
+                    ? 'Пробный период'
+                    : subscriptionStatus === 'active'
+                      ? 'Активна'
+                      : 'Истекла'
+                }</div>
+              </div>
+              <div className="chevron">{userMenuOpen ? '▲' : '▼'}</div>
+            </button>
+            {userMenuOpen && (
+              <div className="user-menu">
+                <Link href="/settings" className="user-menu-item">Настройки</Link>
+                <Link href="/profile" className="user-menu-item">Профиль</Link>
+                <Link href="/help" className="user-menu-item">Помощь</Link>
+                <hr className="user-menu-separator" />
+                <button onClick={handleLogout} className="user-menu-item">Выйти</button>
+              </div>
+            )}
+          </div>
+        </header>
         <div className="content-header">
           <h1 className="section-title">Каталог ИИ-помощников</h1>
