+++ pages/dashboard.tsx
@@ -2,6 +2,8 @@
 import Link from 'next/link';
 import { useEffect, useState } from 'react';
 import Sidebar from '@/components/Sidebar';
+import HamburgerIcon from '@/components/HamburgerIcon';
+import CloseIcon from '@/components/CloseIcon';
 
 
 const popularAgents = [
@@ -39,11 +41,6 @@ export default function Dashboard() {
     setSidebarOpen(window.innerWidth > 768);
   }, []);
 
-  useEffect(() => {
-    const handler = () => setSidebarOpen(prev => !prev);
-    document.addEventListener('toggleSidebar', handler);
-    return () => document.removeEventListener('toggleSidebar', handler);
-  }, []);
 
   useEffect(() => {
     fetch('/api/me', { credentials: 'include' })
@@ -66,6 +63,17 @@ export default function Dashboard() {
     setUserMenuOpen(!userMenuOpen);
   };
 
+  const handleLogout = async () => {
+    try {
+      const res = await fetch('/api/logout', { credentials: 'include' });
+      if (res.ok) {
+        window.location.href = '/auth/login';
+      }
+    } catch (e) {
+      console.error('Ошибка при выходе:', e);
+    }
+  };
+
   return (
     <div className="dashboard-layout">
       {/* Sidebar */}
@@ -78,6 +86,36 @@ export default function Dashboard() {
 
       {/* Main Content */}
       <main className={`main-content ${sidebarOpen ? 'with-sidebar' : 'full-width'}`}>
+        <header className="lk-header">
+          <button className="mobile-hamburger" onClick={toggleSidebar}>
+            {sidebarOpen ? <CloseIcon /> : <HamburgerIcon />}
+          </button>
+          <div className="lk-header__profile">
+            <button className="user-info-button" onClick={toggleUserMenu}>
+              <div className="lk-header__avatar">{email.charAt(0).toUpperCase()}</div>
+              <div className="user-details">
+                <div className="user-email">{email}</div>
+                <div className={`subscription-badge ${subscriptionStatus}`}>{
+                  subscriptionStatus === 'trial'
+                    ? 'Пробный период'
+                    : subscriptionStatus === 'active'
+                      ? 'Активна'
+                      : 'Истекла'
+                }</div>
+              </div>
+              <div className="chevron">{userMenuOpen ? '▲' : '▼'}</div>
+            </button>
+            {userMenuOpen && (
+              <div className="user-menu">
+                <Link href="/settings" className="user-menu-item">Настройки</Link>
+                <Link href="/profile" className="user-menu-item">Профиль</Link>
+                <Link href="/help" className="user-menu-item">Помощь</Link>
+                <hr className="user-menu-separator" />
+                <button onClick={handleLogout} className="user-menu-item">Выйти</button>
+              </div>
+            )}
+          </div>
+        </header>
         <div className="content-header">
           <h2>Добро пожаловать!</h2>
           {(subscriptionStatus === 'expired' || subscriptionStatus === 'trial') && (
